# Conflicts:
#	keep/api/core/db.py
#	keep/api/routes/preset.py

+++ keep/api/models/workflow.py
@@ -100,7 +100,7 @@ class WorkflowExecutionDTO(BaseModel):
     status: str
     logs: Optional[List[WorkflowExecutionLogsDTO]]
     error: Optional[str]
-    execution_time: Optional[int]
+    execution_time: Optional[float]
     results: Optional[dict]
